pyproject.toml
[project]
name = "iokit"
description = "Input Output Kit"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dynamic = ["version"]
authors = [
    {name = "Vladislav A. Proskurov", email = "rilshok@pm.me"},
]
dependencies = [
    "humanize>=4.9.0",
    "jsonlines>=4.0.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.1",
    "pytz>=2024.1",
    "requests>=2.32.3",
    "typing-extensions>=4.8.0",
    "cryptography>=41.0.7",
    "numpy>=1.21.1",
    "soundfile>=0.12.1",
    "pandas>=1.5.3",
    "Pillow>=10.4.0",
    "xxhash>=3.4.1",
]


[tool.setuptools.dynamic]
version = {attr = "iokit.__version__"}


[project.optional-dependencies]
dev = ["iokit[lint,test]"]
lint = [
    "mypy>=1.7.1",
    "ruff>=0.6.3",
    "types-pytz>=2024.1",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
    "types-python-dateutil>=2.8.19",
]
test = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.6.1",
]


[project.urls]
Homepage = "https://github.com/rilshok/iokit"
Repository = "https://github.com/rilshok/iokit.git"
Issues = "https://github.com/rilshok/iokit/issues"


[tool.mypy]
strict = true


[tool.vulture]
make_whitelist = true
sort_by_size = true
verbose = false
min_confidence = 100
paths = ["src/iokit"]


[tool.ruff]
line-length = 100


[tool.ruff.format]
docstring-code-format = true


[tool.ruff.lint]
select = [
    "RUF",
    "W",
    "E",
    "F",
    "UP",
    "YTT",
    "B",
    "C4",
    "T10",
    "SIM",
    "I",
    "PGH",
    "COM",
    "RSE",
    "FLY",
    "RET",
    "TC",
    "FURB",
    "A",
    "S",
    "TRY",
    "EM",
    "ANN",
    "BLE",
    "N",
    "C90",
    "FBT",
    "SLF",
    "TID",
    "ARG",
    "PTH",
    "NPY",
    "PERF",
    "PL",
    "FIX",
    "TD",
]
ignore = [
    "RUF100",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004", "PLR0915", "S101", "S106", "ANN201", "SLF001"]
